Handling errors in admin article create/update.
Article as it is now has no validation, so the controller had left this bit out.
This isn't so good because a) maybe Article should have validation, and b) it shouldn't be put on
the plugin user to add this functionality to the controller simply because they change the model.

Also, this was easily taken from the CMS plugin, as many things are.