Adding admin articles controller to init list.
So the categories plugin can alter the behavior and loading won't get fucked.

Seriously, I hate this.

Also, this sadly means it won't work in app mode because controllers can't be loaded before init
(when plugins load). This might be the beginning of the end for flogiston bits as apps + plugins.